requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
Pillow>=9.0.0
openai>=1.0.0
python-dotenv>=0.20.0
